+++ lib/form/connection_form/connection_form.dart
@@ -174,8 +174,6 @@ class ConnectionFormNotifier extends StateNotifier<ConnectionFormState> {
 
       state = state.copyWith(error: e.toString(), isSubmitting: false);
       return false;
-    } finally {
-      state = state.copyWith(isSubmitting: false);
     }
   }