+++ lib/providers/services/services_provider.dart
@@ -21,7 +21,7 @@ class ServicesNotifier extends StateNotifier<ServicesState> {
   Future<void> _init() async {
     // Listen to client state changes
     _connectionStateSubscription = ref.listen(clientStateProvider(params.alias), (previous, next) {
-      if (next.value == cmk_api.ConnectionState.connected) {
+      if (next.hasValue && next.value == cmk_api.ConnectionState.connected) {
         _startRefreshTimer();
         _fetchData();
       } else {